addin a new branch, ad set a scss and reorganize the app folder structure
