+++ src/app/serviceWorker.js
