+++ src/index.js
@@ -2,9 +2,9 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import firebase from 'firebase';
 
-import './index.css';
-import App from './App';
-import * as serviceWorker from './serviceWorker';
+import './index.scss';
+import App from './app/App';
+import * as serviceWorker from './app/serviceWorker';
 
 // Initialize Firebase
 firebase.initializeApp({
